style(dto 기능별 분류): dto 기능별 폴더구조 수정
board, booking, store 폴더구조 세분화
