+++ src/main/java/com/example/yummytable/domain/Booking.java
@@ -51,5 +51,4 @@ public class Booking {
   private LocalDateTime unregisteredAt;
 
 
-
 }
