+++ src/main/java/com/example/yummytable/dto/board/CreateBoard.java
@@ -1,5 +1,6 @@
-package com.example.yummytable.dto;
+package com.example.yummytable.dto.board;
 
+import com.example.yummytable.dto.board.BoardDto;
 import com.example.yummytable.type.BoardStatus;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.Setter;
 import org.antlr.v4.runtime.misc.NotNull;
 
 
-public class DeleteBoard {
+public class CreateBoard {
 
   @Getter
   @Setter
@@ -19,9 +20,12 @@ public class DeleteBoard {
   @Builder
   public static class Request {
 
-    // 게시글 관련 정보
     @NotNull
     private Long boardId;
+    @NotNull
+    private String title;
+    @NotNull
+    private String content;
     @NotNull
     private String password;
 
@@ -40,19 +44,14 @@ public class DeleteBoard {
     private String content;
     private BoardStatus boardStatus;
     private LocalDateTime registeredAt;
-    private LocalDateTime updatedAt;
-    private LocalDateTime unregisteredAt;
-
 
     public static Response from(BoardDto boardDto) {
       return Response.builder()
           .boardId(boardDto.getBoardId())
           .title(boardDto.getTitle())
           .content(boardDto.getContent())
-          .boardStatus(BoardStatus.DELETE)
+          .boardStatus(BoardStatus.EXISTENT)
           .registeredAt(boardDto.getRegisteredAt())
-          .updatedAt(boardDto.getUpdatedAt())
-          .unregisteredAt(boardDto.getUnregisteredAt())
           .build();
     }
   }
