+++ src/main/java/com/example/yummytable/dto/board/ReadBoard.java
@@ -1,5 +1,6 @@
-package com.example.yummytable.dto;
+package com.example.yummytable.dto.board;
 
+import com.example.yummytable.dto.board.BoardDto;
 import com.example.yummytable.type.BoardStatus;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.Setter;
 import org.antlr.v4.runtime.misc.NotNull;
 
 
-public class CreateBoard {
+public class ReadBoard {
 
   @Getter
   @Setter
@@ -23,24 +24,8 @@ public class CreateBoard {
     @NotNull
     private Long boardId;
     @NotNull
-    private String title;
-    @NotNull
-    private String content;
-    @NotNull
     private String password;
-    private BoardStatus boardStatus;
 
-    // 식당 관련 정보
-    @NotNull
-    private String storeName;
-    private String keyword;
-    @NotNull
-    private Double locationX;
-    @NotNull
-    private Double locationY;
-    private String menu;
-    @NotNull
-    private int capacity;
   }
 
   @Getter
@@ -57,16 +42,6 @@ public class CreateBoard {
     private BoardStatus boardStatus;
     private LocalDateTime registeredAt;
 
-
-    // 식당 관련 정보
-    private String storeName;
-    private String keyword;
-    private Double locationX;
-    private Double locationY;
-    private String menu;
-    private int capacity;
-
-
     public static Response from(BoardDto boardDto) {
       return Response.builder()
           .boardId(boardDto.getBoardId())
@@ -74,12 +49,6 @@ public class CreateBoard {
           .content(boardDto.getContent())
           .boardStatus(BoardStatus.EXISTENT)
           .registeredAt(boardDto.getRegisteredAt())
-          .storeName(boardDto.getStoreName())
-          .keyword(boardDto.getKeyword())
-          .locationX(boardDto.getLocationX())
-          .locationY(boardDto.getLocationY())
-          .menu(boardDto.getMenu())
-          .capacity(boardDto.getCapacity())
           .build();
     }
   }
