+++ src/main/java/com/example/yummytable/dto/booking/ReadBooking.java
@@ -1,19 +1,19 @@
-package com.example.yummytable.dto;
+package com.example.yummytable.dto.booking;
 
 import com.example.yummytable.type.BookingStatus;
 import jakarta.validation.constraints.Pattern;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.antlr.v4.runtime.misc.NotNull;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
 
-public class DeleteBooking {
+public class ReadBooking {
 
   @Getter
   @Setter
@@ -24,8 +24,6 @@ public class DeleteBooking {
 
     @Pattern(regexp = "\\d{4}-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])", message = "yyyy-MM-dd 형태로 입력해주세요")
     private String bookingDate;
-    @NotNull
-    private int numberOfApplicants;
 
   }
 
@@ -43,7 +41,7 @@ public class DeleteBooking {
     private int capacity;
     private int numberOfApplicants;
 
-    private String bookingDate;
+    private LocalDate bookingDate;
     private BookingStatus bookingStatus;
 
     @CreatedDate
@@ -59,8 +57,8 @@ public class DeleteBooking {
           .storeId(bookingDto.getStoreId())
           .capacity(bookingDto.getCapacity())
           .numberOfApplicants(bookingDto.getNumberOfApplicants())
-          .bookingDate(String.valueOf(bookingDto.getBookingDate()))
-          .bookingStatus(BookingStatus.DELETE)
+          .bookingDate(bookingDto.getBookingDate())
+          .bookingStatus(BookingStatus.EXISTENT)
           .registeredAt(bookingDto.getRegisteredAt())
           .updatedAt(bookingDto.getUpdatedAt())
           .unregisteredAt(bookingDto.getUnregisteredAt())
