+++ src/main/java/com/example/yummytable/dto/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.example.yummytable.dto;
+package com.example.yummytable.dto.error;
 
 import com.example.yummytable.type.ErrorCode;
 import lombok.AllArgsConstructor;
