+++ src/main/java/com/example/yummytable/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.example.yummytable.exception;
 
-import com.example.yummytable.dto.ErrorResponse;
+import com.example.yummytable.dto.error.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -12,7 +12,7 @@ public class GlobalExceptionHandler {
   @ExceptionHandler(yummyException.class)
   public ErrorResponse handleBoardException(yummyException e) {
     log.error("{} is occurred.", e.getErrorCode());
-    return new com.example.yummytable.dto.ErrorResponse(e.getErrorCode(), e.getErrorMessage());
+    return new ErrorResponse(e.getErrorCode(), e.getErrorMessage());
   }
 
 }
